+++ library/src/main/java/ru/yandex/money/android/sdk/impl/InMemoryColorSchemeRepository.kt
@@ -1,0 +1,7 @@
+package ru.yandex.money.android.sdk.impl
+
+import ru.yandex.money.android.sdk.ColorScheme
+
+object InMemoryColorSchemeRepository {
+    var colorScheme: ColorScheme = ColorScheme.getDefaultScheme()
+}